requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyMySQL==1.1.0
ofxparse==0.21
python-dotenv==1.0.1
Werkzeug==3.0.2
pypdf==6.10.2
